+++ CHANGELOG.md
@@ -355,3 +355,5 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Custom monster fights (kiooeht)
 * Custom bosses (kiooeht)
 * Fix and improve custom events API (kiooeht)
+* Custom bottle-style relics (kiooeht)
+* Make StartBattleHook work when loading a save (gogo81745)